refactor(ui): extract shared utilities and remove duplicate code
- Created shared utility modules following CMP best practices:
  * ChartUtils.kt: Centralized chart calculation logic (calculateChartPoints, calculatePriceStats, limitKlinesForChart)
  * ThemeUtils.kt: Reusable theme calculation (isDarkTheme)
  * PriceUtils.kt: Price change color determination and gradient generation
  * ShimmerUtils.kt: Reusable shimmer effects for loading states

- Refactored UI components to use shared utilities:
  * CryptoListScreen.kt: Removed duplicate chart/theme/price logic, simplified code
  * CoinDetailScreen.kt: Removed duplicate chart calculation and shimmer effects
  * App.kt: Simplified navigation logic, removed complex refresh trigger
  * SettingsScreen.kt: Replaced duplicate theme calculation with utility

- Removed unused model classes and functions:
  * MarkPriceUpdate: Unused data class removed
  * OrderBookSnapshot: Unused data class removed
  * OrderBookData.getCumulativeDepth(): Unused function removed (calculation done inline)

- Fixed composable invocation issues:
  * Updated isDarkTheme() calls to use composable context correctly
  * Fixed getPriceChangeColor() to accept primaryColor parameter

Benefits:
- Eliminated ~200+ lines of duplicate code
- Single source of truth for chart/theme/price logic
- Improved maintainability and consistency
- Better adherence to CMP best practices
- Cleaner, more organized codebase
