+++ composeApp/src/commonMain/kotlin/ui/utils/ShimmerUtils.kt
@@ -1,0 +1,70 @@
+package ui.utils
+
+import androidx.compose.animation.core.LinearEasing
+import androidx.compose.animation.core.RepeatMode
+import androidx.compose.animation.core.animateFloat
+import androidx.compose.animation.core.infiniteRepeatable
+import androidx.compose.animation.core.rememberInfiniteTransition
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.background
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+
+/**
+ * Shared shimmer effect utilities following CMP best practices.
+ * Extracts reusable shimmer logic to avoid duplication.
+ */
+
+/**
+ * Creates a simple shimmer effect modifier for loading placeholders.
+ */
+@Composable
+fun Modifier.shimmerEffect(): Modifier {
+    val shimmerColors = listOf(
+        Color.LightGray.copy(alpha = 0.6f),
+        Color.LightGray.copy(alpha = 0.3f),
+        Color.LightGray.copy(alpha = 0.6f)
+    )
+
+    return this.background(
+        brush = Brush.horizontalGradient(shimmerColors)
+    )
+}
+
+/**
+ * Creates an animated shimmer effect modifier for loading placeholders.
+ * More sophisticated than the simple shimmer effect.
+ */
+@Composable
+fun Modifier.animatedShimmerEffect(): Modifier {
+    val infiniteTransition = rememberInfiniteTransition(label = "shimmer")
+    val shimmerTranslate by infiniteTransition.animateFloat(
+        initialValue = 0f,
+        targetValue = 1000f,
+        animationSpec = infiniteRepeatable(
+            animation = tween(1000, easing = LinearEasing),
+            repeatMode = RepeatMode.Restart
+        ),
+        label = "shimmer_translate"
+    )
+
+    val shimmerColors = listOf(
+        MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.6f),
+        MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.2f),
+        MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.6f)
+    )
+
+    return this.background(
+        brush = Brush.linearGradient(
+            colors = shimmerColors,
+            start = Offset(shimmerTranslate - 300f, shimmerTranslate - 300f),
+            end = Offset(shimmerTranslate, shimmerTranslate)
+        )
+    )
+}
+
